feat: do client credentials in dev
